+++ android/app/src/main/java/com/miguelespinoza/goku/MainActivity.java
@@ -20,27 +20,4 @@ public class MainActivity extends ReactActivity {
     protected String getMainComponentName() {
         return "goku";
     }
-
-    /**
-     * Returns whether dev mode should be enabled.
-     * This enables e.g. the dev menu.
-     */
-    @Override
-    protected boolean getUseDeveloperSupport() {
-        return BuildConfig.DEBUG;
-    }
-
-    /**
-     * A list of packages used by the app. If the app uses additional views
-     * or modules besides the default ones, add more packages here.
-     */
-    @Override
-    protected List<ReactPackage> getPackages() {
-        return Arrays.<ReactPackage>asList(
-            new MainReactPackage(),
-            new GokuReactPackage(),
-            new ReactMaterialKitPackage(),
-            new RealmReactPackage()
-        );
-    }
 }
