+++ android/app/src/main/java/com/miguelespinoza/goku/MainApplication.java
@@ -1,0 +1,44 @@
+package com.miguelespinoza.goku;
+
+import android.app.Application;
+
+import com.facebook.react.ReactApplication;
+import com.facebook.react.ReactNativeHost;
+import com.facebook.react.ReactPackage;
+import com.facebook.react.shell.MainReactPackage;
+import com.github.xinthink.rnmk.ReactMaterialKitPackage;
+
+import java.util.Arrays;
+import java.util.List;
+
+import io.realm.react.RealmReactPackage;
+
+public class MainApplication extends Application implements ReactApplication {
+
+    private final ReactNativeHost mReactNativeHost = new ReactNativeHost(this) {
+        @Override
+        public boolean getUseDeveloperSupport() {
+            return BuildConfig.DEBUG;
+        }
+
+        @Override
+        protected List<ReactPackage> getPackages() {
+            return Arrays.<ReactPackage>asList(
+                    new MainReactPackage(),
+                    new GokuReactPackage(),
+                    new ReactMaterialKitPackage(),
+                    new RealmReactPackage()
+            );
+        }
+    };
+
+    @Override
+    public ReactNativeHost getReactNativeHost() {
+        return mReactNativeHost;
+    }
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+    }
+}
